requirements.txt
streamlit>=1.35.0
anthropic>=0.34.0
python-docx>=1.1.2
mammoth>=1.8.0
pypdf>=4.3.1
pandas>=2.2.2
openpyxl>=3.1.5
python-dotenv>=1.0.1
